task 14: fix bug
