+++ flutterapp/README.md
@@ -6,6 +6,4 @@
 ![Screenshot](simulator_screenshot_5BA23260-4F40-4901-81CA-5042DF9A5970.png)
 ![Screenshot](simulator_screenshot_82FEF916-62EC-4252-9F91-EFB4933594F0.png)
 
-## Problems
-Не получилось заблюрить фотографию справа
