+++ flutterapp/lib/14-task/utils/ccolors.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+abstract class CColors {
+  static Color primary = const Color(0xFF222222);
+  static Color greyText = const Color(0xFFBFBFBF);
+}
