+++ flutterapp/lib/14-task/utils/images.dart
@@ -1,23 +1,4 @@
 abstract class Images {
   static String logo = 'assets/images/14-task/logo.png';
-  static String photo_1 = 'assets/images/14-task/photo_1.png';
-  static String photo_2 = 'assets/images/14-task/photo_2.png';
-  static String photo_3 = 'assets/images/14-task/photo_3.png';
-  static String photo_4 = 'assets/images/14-task/photo_4.png';
-  static String photo_5 = 'assets/images/14-task/photo_5.png';
-  static String photo_6 = 'assets/images/14-task/photo_6.png';
-  static String photo_7 = 'assets/images/14-task/photo_7.png';
-  static String photo_8 = 'assets/images/14-task/photo_8.png';
-  static String photo_9 = 'assets/images/14-task/photo_9.png';
-  static String photo_10 = 'assets/images/14-task/photo_10.png';
-  static String photo_11 = 'assets/images/14-task/photo_11.png';
-  static String photo_12 = 'assets/images/14-task/photo_12.png';
-  static String photo_13 = 'assets/images/14-task/photo_13.png';
-  static String photo_14 = 'assets/images/14-task/photo_14.png';
-  static String photo_15 = 'assets/images/14-task/photo_15.png';
-  static String photo_16 = 'assets/images/14-task/photo_16.png';
-  static String photo_17 = 'assets/images/14-task/photo_17.png';
-  static String photo_18 = 'assets/images/14-task/photo_18.png';
-  static String photo_19 = 'assets/images/14-task/photo_19.png';
-
+  static List<String> photos = List.generate(19, (index) => 'assets/images/14-task/photo_${index+1}.png');
 }
